refactor: split metadata form into two files
so it will be easier to merge in ee
